requirements.txt
streamlit==1.31.1
streamlit-extras==0.4.0
streamlit-vertical-slider==2.5.5
easyocr==1.6.2
numpy==1.24.2
opencv-python-headless==4.5.4.60
Pillow==9.5.0


